% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Message de bienvenue!}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{a string}
}
\value{
a string
}
\description{
Message de bienvenue!
}
\examples{
dire_bonjour()
## par defaut la fonction dire_bonjour retourne "Bonjour, toi"
dire_bonjour(prenom="untel")
## retourne "Bonjour, untel"

}
